requirements.txt
notion-client>=2.2.1
Typer>=0.12.5
pydantic>=2.8.2
python-dateutil>=2.9.0
rich>=13.9.2
feedparser>=6.0.11
trafilatura>=1.9.0
youtube-transcript-api>=0.6.2
tenacity>=9.0.0
PyYAML>=6.0.2
openai>=1.0.0
tiktoken>=0.5.0
requests>=2.31.0
yt-dlp>=2024.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

